changed file names ver 0.1
Changes:

Changed Names of files to better suit the library.
Moved the units creation on the y scale to a seperate function of the
Graph class called "units".
An examples directory has been created, containing a basic example.
A bug reelated to the calculation of the units on the y scale not
relating to the maxY property was fixed.

Bugs:

The bar charts are not animated when they are drawn on the graph, even
though a transition is set on their height and width properties.